Perform query with unmodified input string.
Now that the search engine uses shingles, it doesn't make sense to turn
the query into a set of keywords. The question, as entered, is now sent
to ES in the query. Also no longer using extracted concepts/entities in
the query (as this is already done at index time). Data is still
retrieved for the purpose of highlighting concepts to the user, but they
aren't used in the query.

Also moved config to a separate file so code doesn't need to be
modified.