Implement Dockerfile
Remove use of assembly plugin - just packaging the project as a regular
Jar now and passing the dependencies into the -cp param of the java
command. This reduces overall image size as the deps are already
downloaded when building, so there's no need to package them up into the
jar as well. Also had to modify the HTTP server config so it wasn't
hardcoded to localhost, but instead 0.0.0.0.