Update config to use DNS alias for ES.
This is being done in anticipation of configuring a number of services
to run using docker compose. This will allow the app to connect to
elasticsearch without requiring any config change. Users not running the
application using our docker compose config are still free to modify the
application settings to suit their needs.