Update docker configuration to enable passing args
Reverted the default elasticsearch host back to localhost, as this is a
better default when not using Docker to run the project. Set up
build-time arguments to allow setting the meaning cloud API key and the
elastic search hostname & port when building the image. These values can
be passed in by docker-compose.